requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
whoosh>=2.7.4
numpy>=1.26.0
google-genai>=0.3.0
tqdm>=4.66.0
beautifulsoup4>=4.12.2
ebooklib>=0.18
unidecode>=1.3.7
nltk>=3.9
